Add bug fixes, code refactoring

+++ library/TelAPI.Api/Model/VoiceEffectOptions.cs
@@ -5,9 +5,9 @@ namespace TelAPI
     public class VoiceEffectOptions
     {
         public AudioDirection? AudioDirection { get; set; }
-        public int? Pitch { get; set; }
-        public int? PitchSemiTones { get; set; }
-        public int? PitchOctaves { get; set; }
-        public int? Rate { get; set; }
+        public double? Pitch { get; set; }
+        public double? PitchSemiTones { get; set; }
+        public double? PitchOctaves { get; set; }
+        public double? Rate { get; set; }
     }
 }
